Prep work for getting control and driving to work
Removed redudant things in screen.hpp, this is going to be incorporated
into a seperate file. Started beginning work for Auton and Driving
controls in file, will work on styling buttons later.

Note: I know the buttons and everything in that tab is broken also, will
be fixed before comp